Re-implementing natural stone calculation. DIKERNEL-56

+++ DiKErnel/Src/DiKErnel.FunctionLibrary/Public/Revetment.cpp
@@ -28,9 +28,9 @@ namespace DiKErnel::FunctionLibrary
         return numberOfTimeStamps - 1;
     }
 
-    double Revetment::IncrementTime(
-        const double startTime,
-        const double endTime)
+    int Revetment::IncrementTime(
+        const int startTime,
+        const int endTime)
     {
         return endTime - startTime;
     }
